requirements.txt
altgraph==0.17.4
fpdf==1.7.2
langdetect==1.0.9
lxml==5.2.2
packaging==24.1
pdf2image==1.17.0
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pytesseract==0.3.10
python-docx==1.1.2
pywin32-ctypes==0.2.2
six==1.16.0
typing_extensions==4.12.2
